Reorganize OpenID config
- Split into multiple files

- Remove JSON unmarshaler for Config and providerCfg types (unused)
